backend/resume-analysis-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
docx2txt==0.8
pdfminer.six==20231228
language-tool-python==2.7.1
google-cloud-firestore==2.14.0
